btrfs-progs: restore: Redirect looping prompt to stderr
When initiating a btrfs restore, i.e. in verbose mode, it might be
helpful to redirect stdout to a log file.  stdout redirection to a file
is usually fully buffered, so the looping prompt is likely not flushed
to the log file for the user to notice.

Thus, btrfs restore might seem to hang, while it is actually waiting for
a user interaction.

To improve the user experience, send the looping prompt to (unbuffered)
stderr.

Signed-off-by: Raoul Bhatia <raoul@bhatia.at>